Rename combinator module to builder
