+++ examples/animation/banner_bounce.rs
@@ -1,13 +1,13 @@
 use std::f32::consts::PI;
 
 use bevy::{
-    color::{Srgba, palettes::css::WHITE},
+    color::{palettes::css::WHITE, Srgba},
     core_pipeline::{bloom::BloomSettings, tonemapping::Tonemapping},
     prelude::*,
     window,
 };
 use bevy_tween::{
-    combinator::{go, parallel, tween_exact, AnimationCommands},
+    builder::{go, parallel, tween_exact, AnimationCommands},
     prelude::*,
 };
