+++ examples/animation/banner_triangle.rs
@@ -2,7 +2,7 @@ use std::f32::consts::TAU;
 
 use bevy::prelude::*;
 use bevy_tween::{
-    combinator::{parallel, tween_exact, AnimationCommands},
+    builder::{parallel, tween_exact, AnimationCommands},
     interpolate::{angle_z, translation},
     prelude::*,
     tween::{AnimationTarget, TargetComponent},
